fix: change api userprofile path and use userid from request instead of decode token
